Added search by regex string in top bar
